amplify/agent/runtime/requirements.txt
bedrock-agentcore
strands-agents[otel]
aws-opentelemetry-distro
